int -> long
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@261831 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/math.c
@@ -1068,7 +1068,7 @@ PHP_FUNCTION(number_format)
 {
 	zval *sep1 = NULL, *sep2 = NULL;
 	double num;
-	int dec = 0;
+	long dec = 0;
 	char thousand_sep=',', dec_point='.';
 	char *tmp;
